// src/store/types.ts
export enum ActionTypes {
  LOADING = "LOADING",
  ERROR = "ERROR",
  CHANGE_LANGUAGE = "CHANGE_LANGUAGE",
  FETCHED_REPOS = "FETCHED_REPOS",
  FETCHED_FOLDER_CONTENT = "FETCHED_FOLDER_CONTENT",
  LOAD_FOLDER_CONTENT = "LOAD_FOLDER_CONTENT",
  OPEN_README = "OPEN_README",
  OPEN_IMAGE = "OPEN_IMAGE",
  OPEN_START_MENU = "OPEN_START_MENU",
  OPEN_WINDOW = "OPEN_WINDOW",
  CLOSE_WINDOW = "CLOSE_WINDOW",
  OPEN_ALERT = "OPEN_ALERT",
  CLOSE_ALERT = "CLOSE_ALERT",
  SET_ACTIVE_WINDOW = "SET_ACTIVE_WINDOW"
}